+++ src/wifi/model/eht/advanced-emlsr-manager.cc
@@ -63,7 +63,7 @@ class EmlsrPhyListener : public WifiPhyListener
                             m_emlsrManager);
     }
 
-    void NotifyRxEndError() override
+    void NotifyRxEndError(const WifiTxVector& /* txVector */) override
     {
     }