+++ src/wifi/test/spectrum-wifi-phy-test.cc
@@ -327,9 +327,9 @@ class TestPhyListener : public ns3::WifiPhyListener
         ++m_notifyRxEndOk;
     }
 
-    void NotifyRxEndError() override
+    void NotifyRxEndError(const WifiTxVector& txVector) override
     {
-        NS_LOG_FUNCTION(this);
+        NS_LOG_FUNCTION(this << txVector);
         ++m_notifyRxEndError;
     }